+++ lib/catlixir/command/help.ex
@@ -13,7 +13,7 @@ defmodule Catlixir.Command.Help do
     :ok
   end
 
-  @command Application.get_env(:catlixir, :command)
+  @command Catlixir.get_command_prefix!()
 
   @doc """
   Generates a nostrum embed which contains all of the commands which
@@ -23,16 +23,11 @@ defmodule Catlixir.Command.Help do
     import Nostrum.Struct.Embed
     import Catlixir.Helper
 
-    me = Nostrum.Cache.Me.get()
-
-    avatar_url =
-      "https://cdn.discordapp.com/avatars/#{me.id}/#{me.avatar}.png"
-
     {:ok, vsn} = :application.get_key(:catlixir, :vsn)
 
     %Nostrum.Struct.Embed{}
-    |> put_title("Co-meow-nds for: #{me.username}!")
-    |> put_thumbnail(avatar_url)
+    |> put_title("Co-meow-nds for: #{Catlixir.get_username!()}!")
+    |> put_thumbnail(Catlixir.get_avatar_url!())
     |> put_field("#{@command} fact", "Get a random fact about us (cats)!")
     |> put_field("#{@command} breed (name)", "Get info about a breed. If the name is not specified, it will return a random breed.")
     |> put_field("#{@command} random", "Get a random image from the `r/cat` subreddit!")