+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 0.1.4
+- Removed `Catlixir.Command.Random` module. 
+- Started using `Catlixir` module.
+
 # 0.1.3
 - Using `RedditPost` module to get memes and random images.